Rename parser/parser.ts to parser/index.ts
