+++ src/cli/index.ts
@@ -3,7 +3,7 @@ import * as fs from 'fs'
 import * as path from 'path'
 import glob from 'tiny-glob'
 import * as _yargs from 'yargs'
-import { parse } from '../parser/parser'
+import { parse } from '../parser'
 import { createDocument } from '../parser/document-builder'
 import { canonicalPath } from '../utils'
 import { Diagnostic } from '../diagnostic'
